new realloc/calloc and some fixes (with comments)

+++ utils.h
@@ -6,7 +6,7 @@
 
 typedef long long Hash_t;
 
-Hash_t hash_func         (void* point, int size); // const void* data - ded wrote
+Hash_t hash_func (void* point, size_t size); // const void* data - ded wrote
 
 // Что-то одно надо оставить
 // void*  poison_realloc    (void* point, size_t old_size, size_t new_size);
